switch packaging from .yaml to .spec
The .spec file is now the the which gets edited. Besides being more
aligned with current practices in Tizen, this is also potentially more
flexible.

The goal was to mark the COPYING file as %license, as indicated by
rpmlint. But it did not become entirely clear to me how rpmbuild picks
up the doc files in the first place (part of lang-find perhaps?), so
simply adding "%license COPYING" was not enough and led to errors
about that file being a duplicate. Therefore I have left packaging as it
was, without fixing the license file warning.